schily-2011-05-08.tar.bz2

+++ patch/util.c
@@ -1,7 +1,7 @@
-/* @(#)util.c	1.14 11/04/22 2011 J. Schilling */
+/* @(#)util.c	1.15 11/05/06 2011 J. Schilling */
 #ifndef lint
 static	char sccsid[] =
-	"@(#)util.c	1.14 11/04/22 2011 J. Schilling";
+	"@(#)util.c	1.15 11/05/06 2011 J. Schilling";
 #endif
 /*
  *	Copyright (c) 1986 Larry Wall
@@ -17,6 +17,10 @@ static	char sccsid[] =
 #define	EXT
 #include "util.h"
 #include <schily/varargs.h>
+#include <schily/errno.h>
+
+static	Llong	fetchtime	__PR((char *t));
+static	Llong	mkgmtime	__PR((struct tm *tp));
 
 /* Rename a file, copying it if necessary. */
 
@@ -32,7 +36,7 @@ move_file(from, to)
 
 	/* to stdout? */
 
-	if (strEQ(to, "-")) {
+	if (from && strEQ(to, "-")) {
 #ifdef DEBUGGING
 		if (debug & 4)
 			say(_("Moving %s to stdout.\n"), from);
@@ -67,9 +71,12 @@ move_file(from, to)
 				simplename = s+1;
 		}
 		/* find a backup name that is not the same file */
+		file_stat.st_ctime = (time_t)0;
 		while (stat(bakname, &file_stat) >= 0 &&
 			to_device == file_stat.st_dev &&
 			to_inode == file_stat.st_ino) {
+			if (file_stat.st_ctime >= starttime) /* mult patches */
+				goto backup_done;
 			for (s = simplename; *s && !islower(*s); s++) {
 				;
 				/* LINTED */
@@ -79,6 +86,8 @@ move_file(from, to)
 			else
 				Strcpy(simplename, simplename+1);
 		}
+		if (file_stat.st_ctime >= starttime) /* mult patches */
+			goto backup_done;
 		while (unlink(bakname) >= 0) {	/* while() is for Eunice */
 			;
 			/* LINTED */
@@ -96,6 +105,16 @@ move_file(from, to)
 			;
 			/* LINTED */
 		}
+	}
+backup_done:
+	if (from == NULL) {
+		if (!do_backup) {
+			if (debug & 4)
+				say(_("Removing file %s\n"), to);
+			if (unlink(to) != 0 && errno != ENOENT)
+				fatal(_("Can't remove file %s\n"), to);
+		}
+		return (0);
 	}
 #ifdef DEBUGGING
 	if (debug & 4)
@@ -123,6 +142,27 @@ move_file(from, to)
 	}
 	Unlink(from);
 	return (0);
+}
+
+void
+removedirs(path)
+	char	*path;
+{
+	char	*p = strrchr(path, '/');
+
+	if (p == NULL)
+		return;
+	while (p > path) {
+		*p = '\0';
+		if (rmdir(path) == 0 && verbose)
+			say(_("Removed empty directory %s\n"),
+				path);
+		*p = '/';
+		while (--p > path) {
+			if (*p == '/')
+				break;
+		}
+	}
 }
 
 /* Copy a file. */
@@ -347,10 +387,10 @@ makedirs(filename, striplast)
 		return;
 #if 1
 	for (i = 0; i <= dirvp; i++) {
-		mkdir(tmpbuf, 
-			S_IRUSR|S_IWUSR|S_IXUSR 
-			|S_IRGRP|S_IWGRP|S_IXGRP 
-			|S_IROTH|S_IWOTH|S_IXOTH); 
+		mkdir(tmpbuf,
+			S_IRUSR|S_IWUSR|S_IXUSR
+			|S_IRGRP|S_IWGRP|S_IXGRP
+			|S_IROTH|S_IWOTH|S_IXOTH);
 		*dirv[i] = '/';
 	}
 #else
@@ -369,15 +409,19 @@ makedirs(filename, striplast)
 /* Make filenames more reasonable. */
 
 char *
-fetchname(at, strip_leading, assume_exists)
+fetchname(at, strip_leading, assume_exists, isnulldate)
 	char *at;
 	int strip_leading;
 	int assume_exists;
+	bool *isnulldate;
 {
 	char *s;
 	char *name;
 	char *t;
 	char tmpbuf[200];
+
+	if (isnulldate)
+		*isnulldate = FALSE;
 
 	if (!at)
 		return (Nullch);
@@ -392,13 +436,18 @@ fetchname(at, strip_leading, assume_exists)
 		say(_("fetchname %s %d %d\n"),
 			name, strip_leading, assume_exists);
 #endif
-	if (strnEQ(name, "/dev/null", 9)) /* files can be created by diffing */
-		return (Nullch);	  /* against /dev/null. */
 	for (; *t && !isspace(*t); t++)
 		if (*t == '/')
 			if (--strip_leading >= 0)
 				name = t+1;
 	*t = '\0';
+	if (strEQ(name, "/dev/null")) {	/* files can be created by diffing */
+		if (isnulldate)		/* against /dev/null. */
+			*isnulldate = TRUE;
+		return (Nullch);
+	}
+	if (isnulldate)
+		*isnulldate = fetchtime(t) == 0;
 	if (name != s && *s != '/') {
 		name[-1] = '\0';
 		if (stat(s, &file_stat) && file_stat.st_mode & S_IFDIR) {
@@ -422,4 +471,97 @@ fetchname(at, strip_leading, assume_exists)
 		}
 	}
 	return (name);
+}
+
+static Llong
+fetchtime(t)
+	char	*t;
+{
+	Llong	d = -1;
+	struct tm tm;
+	int	n;
+	int	h;
+	int	m;
+
+	for (++t; *t != '\0' && isspace(*t); t++)
+		;
+	n = sscanf(t, "%4d-%2d-%2d %2d:%2d:%2d",
+			&tm.tm_year, &tm.tm_mon, &tm.tm_mday,
+			&tm.tm_hour, &tm.tm_min, &tm.tm_sec);
+	tm.tm_year -= 1900;
+	tm.tm_mon -= 1;
+
+	if (n != 6)
+		return (d);
+
+	if (strlen(t) <= 19)
+		return (d);
+
+	t += 19;
+	for (; *t != '\0' && !isspace(*t); t++)
+		;
+	for (; *t != '\0' && isspace(*t); t++)
+		;
+	n = sscanf(t, "%3d%2d", &h, &m);
+	if (n != 2)
+		return (d);
+	if (h < -24 || h > 25)
+		return (d);
+	if (m < 0 || m > 59)
+		return (d);
+	d = mkgmtime(&tm);
+	m += 60 * h;
+	m *= 60;
+	d -= m;
+	return (d);
+}
+
+#define	dysize(A) (((A)%4)? 365 : (((A)%100) == 0 && ((A)%400)) ? 365 : 366)
+/*
+ * Return the number of leap years since 0 AD assuming that the Gregorian
+ * calendar applies to all years.
+ */
+#define	LEAPS(Y) 	((Y) / 4 - (Y) / 100 + (Y) / 400)
+/*
+ * Return the number of days since 0 AD
+ */
+#define	YRDAYS(Y)	(((Y) * 365L) + LEAPS(Y))
+/*
+ * Return the number of days between Januar 1 1970 and the end of the year
+ * before the the year used as argument.
+ */
+#define	DAYS_SINCE_70(Y) (YRDAYS((Y)-1) - YRDAYS(1970-1))
+
+char *Datep;
+static int dmsize[12] = {31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31};
+
+static int
+mosize(y, t)
+int y, t;
+{
+
+	if (t == 2 && dysize(y) == 366)
+		return (29);
+	return (dmsize[t-1]);
+}
+
+static Llong
+mkgmtime(tp)
+	struct tm	*tp;
+{
+	Llong	tim = (time_t)0L;
+	int	y = tp->tm_year + 1900;
+	int	t = tp->tm_mon + 1;
+
+	tim = DAYS_SINCE_70(y);
+	while (--t)
+		tim += mosize(y, t);
+	tim += tp->tm_mday - 1;
+	tim *= 24;
+	tim += tp->tm_hour;
+	tim *= 60;
+	tim += tp->tm_min;
+	tim *= 60;
+	tim += tp->tm_sec;
+	return (tim);
 }
